Safari-611.1.10/LayoutTests/fullscreen/video-inside-flex-item-expected.txt


EVENT(webkitfullscreenchange)
TEST(video.clientWidth == document.body.clientWidth) OK
EVENT(webkitfullscreenchange)
EXPECTED (320 == '320') OK
EXPECTED (240 == '240') OK
END OF TEST
